requirements.txt
Django==4.2.3
jieba==0.42.1
py2neo==2021.2.3
pyahocorasick==1.4.2
PyYAML==6.0
channels==4.0.0
django-sesame==3.1
daphne==4.0.0
mysqlclient==2.2.0
pymysql==1.1.0
torch==2.0.1
text2vec==1.2.2
zhipuai==1.0.7